Rename <data>…</data> block to <get>…</get>
Data is a reserved word (tag) in HTML. Using <get> also mirrors the functionality of GET routes (as this is basically an overloaded GET data route on the same path as the rest of the page). It also means that we can include <post>, <put>, <delete>, etc., in the future and remain consistent.